Change back socket send and receive function signatures to original.

+++ AS/include/AS/Socket.h
@@ -43,9 +43,9 @@ class Socket
       APESEARCH::unique_ptr<Socket> accept(struct sockaddr *addr, socklen_t *addrlen);
       
 
-      virtual ssize_t send(const char* buffer, size_t length);
+      virtual ssize_t send(const char* buffer, int length);
 
-      virtual ssize_t receive(char *buffer, size_t length);
+      virtual ssize_t receive(char *buffer, int length);
 
       int getFD() const 
          {
